jitter: fix signed, and unsigned, mixup
Fixes warnings on 32-bit RasPi

+++ ntpfrob/jitter.c
@@ -44,7 +44,7 @@ get_clocktime(
 	 * Convert Unix clock from seconds and nanoseconds to seconds.
 	 */
 	clock_gettime(CLOCK_REALTIME, &ts);
-	*now = lfptouint(ts.tv_sec + JAN_1970);  /* no fraction, yet */
+	*now = lfptouint(ts.tv_sec + (long)JAN_1970);  /* no fraction, yet */
 	dtemp = ts.tv_nsec * S_PER_NS;
 
 	/*
